[version] Bump version to 40~1
From now on, the versioning scheme will be as follows:

xBoot-AA~BB

Where AA is the total number of commits to this source tree
at the start of the day. Any commits that follow on the same day
won't count towards AA and will be assigned to BB starting at 1.

For Example, the version string:

	xBoot-40~1

Means that for my first commit of the day, the total
number of commits made to this repository is 40. When I
go ahead and push my first commit of the day, BB will be set
to 1, for my second commit, 2, my third commit, 3, and so on.