[lib][device] Implement NVRAM attributes
NVRAM variables now support a set of attributes.

Currently supported attributes are:

    * nv_attr_u    - Unmodofied NVRAM variable
    * nv_attr_m    - Modified NVRAM variable
    * nv_attr_p    - Protected NVRAM variable

When variables are created via setenv(), they are
assigned the default attribute nv_attr_u and will
show up without a flag if listed with printenv.

When variables are updated via setenv(), they are
assigned the attribute nv_attr_m and will show up
with an 'M' flag if listed with printenv.

When variables are created via setenv_protect(), they
are assigned the attribute nv_attr_p and will show
up with a 'P' flag if listed wuth printenv.

NVRAM variables that are created via setenv_protect()
are read-only and can only be set once.