[lib][crt] Make setenv() more standards-compliant
The standard behavior for passing NULL as an envval
to setenv is to delete the variable from the environment.
It is also expected for setenv to fail if the '=' character
is in the envname.

The setenv shell command has also been updated to reflect this behavior.

Also added a sanity check on variable setting length when updating
a variable in the NVRAM layer.